엔빵 페이지 컴포넌트화 및 임시저장
