+++ src/app/community/main/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import DetailHeader from '@/components/common/header/DetailHeader'
-import PostBoard from '@/components/community/PostBoard'
+import PostBoard from '@/components/community/NbreadTabs'
 const CommunityPage = () => {
   return (
     <div className="flex flex-col justify-between p-24">
